merge (#20)
* zstd_compression: Make use of std::span in interfaces

Allows condensing the data and size parameters into a single argument.

* lz4_compression: Make use of std::span in interfaces

Allows compressing the data and size parameters into one.

* system_control: Make functions internally linked where applicable

These functions are only ever used internally as implementation details
for GenerateRandomRange(), so these can be given internal linkage.

* cheat_engine: Resolve implicit bool->u64 conversion

We can just return zero here.

* textures/decoders: Fix block linear to pitch copies

There were two issues with block linear copies. First the swizzling was
wrong and this commit reimplements them.

The other issue was that these copies are generally used to download
render targets from the GPU and yuzu was not downloading them from
host GPU memory unless the extreme GPU accuracy setting was selected.
This commit enables cached memory reads for all accuracy levels.

- Fixes level thumbnails in Super Mario Maker 2.

* kernel/scheduler: Mark SchedulerLock constructor as nodiscard

Allows the compiler to warn about cases where the constructor is used
but then immediately discarded, which is a potential cause of
locking/unlocking bugs.

* emu_window: Mark Scoped constructor and Acquire() as nodiscard

Ensures that callers make use of the constructor, preventing bugs from
silently occurring.

* core: Resolve several -Wextra-semi warnings

We can amend one of the cascade macros to require semicolons in order to
compile. In other cases, we can just remove the superfluous semicolons.

* software_keyboard: Resolve a pessimizing move warning

A std::vector created in place like this is already an rvalue and
doesn't need to be moved.

* CMakeLists: Update fmt to 7.0.3

Keeps the library up to date and fixes a few bugs

* time_zone_content_manager: Collapse auto and default case

Prevents a useless self-assignment from occurring.

Co-authored-by: Lioncash <mathew1800@gmail.com>
Co-authored-by: ReinUsesLisp <reinuseslisp@airmail.cc>
Co-authored-by: bunnei <bunneidev@gmail.com>