merge (#21)
* zstd_compression: Make use of std::span in interfaces

Allows condensing the data and size parameters into a single argument.

* lz4_compression: Make use of std::span in interfaces

Allows compressing the data and size parameters into one.

* system_control: Make functions internally linked where applicable

These functions are only ever used internally as implementation details
for GenerateRandomRange(), so these can be given internal linkage.

* aes_util: Make use of non-template variant of Transcode

Same behavior, less template instantiations.

* partition_data_manager: Eliminate magic value

We can use sizeof to make it obvious at the call site where the value is
coming from.

* cheat_engine: Resolve implicit bool->u64 conversion

We can just return zero here.

* textures/decoders: Fix block linear to pitch copies

There were two issues with block linear copies. First the swizzling was
wrong and this commit reimplements them.

The other issue was that these copies are generally used to download
render targets from the GPU and yuzu was not downloading them from
host GPU memory unless the extreme GPU accuracy setting was selected.
This commit enables cached memory reads for all accuracy levels.

- Fixes level thumbnails in Super Mario Maker 2.

* kernel/scheduler: Mark SchedulerLock constructor as nodiscard

Allows the compiler to warn about cases where the constructor is used
but then immediately discarded, which is a potential cause of
locking/unlocking bugs.

* emu_window: Mark Scoped constructor and Acquire() as nodiscard

Ensures that callers make use of the constructor, preventing bugs from
silently occurring.

* maxwell_3d: Resolve -Wextra-semi warning

Semicolons after a function definition aren't necessary.

* async_shaders: Resolve -Wpessimizing-move warning

Prevents pessimization of the move constructor (which thankfully didn't
actually happen in practice here, given std::thread isn't copyable).

* core: Resolve several -Wextra-semi warnings

We can amend one of the cascade macros to require semicolons in order to
compile. In other cases, we can just remove the superfluous semicolons.

* software_keyboard: Resolve a pessimizing move warning

A std::vector created in place like this is already an rvalue and
doesn't need to be moved.

* CMakeLists: Update fmt to 7.0.3

Keeps the library up to date and fixes a few bugs

* time_zone_content_manager: Collapse auto and default case

Prevents a useless self-assignment from occurring.

* yuzu: Make use of qOverload where applicable

Eliminates a verbose function cast.

* common: Make use of [[nodiscard]] where applicable

Now that clang-format makes [[nodiscard]] attributes format sensibly, we
can apply them to several functions within the common library to allow
the compiler to complain about any misuses of the functions.

* common/compression: Roll back std::span changes

Seems like all compilers don't support std::span yet.

* main: Add an option to modify the currrent game's configuration

Creates a new entry in the Emulation menu called "Configure Current Game..." that is only available if a game is currently being executed in yuzu. When selected, it opens the game properties dialog for the current game.

Thanks to @BSoDGamingYT for reminding me to do this.

* yuzu: Resolve -Wextra-semi warnings

While we're in the same area, we can ensure GameDir member variables are
always initialized to consistent values.

* common/time_zone: Simplify GetOsTimeZoneOffset()

We can simplify this function down into a single line with the use of
fmt. A benefit with the fmt approach is that the fmt variant of
localtime is thread-safe as well, making GetOsTimeZoneOffset()
thread-safe as well.

* configure_hotkeys: Don't translate empty strings

There's no need to translate an empty string. This just gives
translators unnecessary work.

* Vk Async pipeline compilation

* Address feedback. Bruteforce delete duplicates

* Vk Async Worker directly emplace in cache

* Address feedback, add shader compile notifier, update setting text

* move thread 1/4 count computation into allocate workers method

* Morph: Update worker allocation comment

Co-authored-by: Morph <39850852+Morph1984@users.noreply.github.com>

* Remove unneeded newlines, optional Registry in shader params

Addressing feedback from Rodrigo

Co-authored-by: Lioncash <mathew1800@gmail.com>
Co-authored-by: ReinUsesLisp <reinuseslisp@airmail.cc>
Co-authored-by: bunnei <bunneidev@gmail.com>
Co-authored-by: lat9nq <lat9nq@virginia.edu>
Co-authored-by: ameerj <aj662@drexel.edu>
Co-authored-by: Ameer J <52414509+ameerj@users.noreply.github.com>
Co-authored-by: Morph <39850852+Morph1984@users.noreply.github.com>
Co-authored-by: David <25727384+ogniK5377@users.noreply.github.com>